Move `gen_wrappers` to codegen module
